+++ app/domains/beratungshilfe/services/pdf/sections/F_besitz/__test__/F_besitz.test.ts
@@ -1,12 +1,12 @@
 import { type BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
-import { getBeratungshilfeParameters } from "~/domains/beratungshilfe/pdf";
-import {
-  fillFinancialBankkonto,
-  fillFinancialGrundeigentum,
-} from "~/domains/beratungshilfe/pdf/sections/F_besitz/F_besitz";
 import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import { fillVermoegenswerte } from "../fillVermoegenswerte";
+import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
+import {
+  fillFinancialBankkonto,
+  fillFinancialGrundeigentum,
+} from "../F_besitz";
 
 describe("F_besitz", () => {
   describe("fillFinancialBankkonto", () => {
