+++ app/domains/beratungshilfe/services/pdf/sections/G_ausgaben.ts
