+++ app/domains/beratungshilfe/services/pdf/sections/__test__/G_ausgaben.test.ts
@@ -1,11 +1,8 @@
 import { CheckboxValue } from "~/components/inputs/Checkbox";
 import { type BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
-import { getBeratungshilfeParameters } from "~/domains/beratungshilfe/pdf";
-import {
-  AUSGABEN_ATTACHMENT_TITLE,
-  fillAusgaben,
-} from "~/domains/beratungshilfe/pdf/sections/G_ausgaben";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
+import { AUSGABEN_ATTACHMENT_TITLE, fillAusgaben } from "../G_ausgaben";
+import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
 
 const expensesRecipients = "ausgaben empfänger";
 const paymentDeadlineDate = "12.12.2099";
