+++ app/domains/beratungshilfe/services/pdf/sections/footer.ts
